Extract detail page navigation logic to navbar
