+++ packages/ui/src/components/domain/details/Navbar.tsx
@@ -1,24 +1,80 @@
+import { useState, useEffect } from 'react'
 import { FiX, FiChevronUp, FiChevronDown } from 'react-icons/fi'
 import Link from 'next/link'
 import Button from '@/components/shared/Button'
-import { useLastViewedPageState } from '@/hooks/useProjectTableState'
+import { useProjectIdsQuery, Project } from '@/graphql/generated/gql'
+import {
+  useBookmarkedProjectsState,
+  useCategoryProjectsState,
+  useLastViewedPageState,
+  useTrendingProjectsState
+} from '@/hooks/useProjectTableState'
+import { defaultSort } from '../projects/types'
 
 type NavbarProps = {
-  currentProjectIndex?: number
-  nextProjectId?: string
-  previousProjectId?: string
-  projectsLength?: number
+  id?: string
   loading?: boolean
 }
 
-const Navbar = ({
-  currentProjectIndex,
-  nextProjectId,
-  previousProjectId,
-  projectsLength,
-  loading
-}: NavbarProps) => {
+const Navbar = ({ id, loading }: NavbarProps) => {
+  const { filters: trendingFilters, sorting: trendingSorting } = useTrendingProjectsState()
+  const { filters: bookmarkFilters, sorting: bookmarkSorting } = useBookmarkedProjectsState()
+  const { filters: categoryFilters, sorting: categorySorting } = useCategoryProjectsState()
   const { lastViewedPage } = useLastViewedPageState()
+
+  const filters =
+    lastViewedPage === 'trending'
+      ? trendingFilters
+      : lastViewedPage === 'bookmarked'
+        ? bookmarkFilters
+        : categoryFilters
+
+  const sorting =
+    lastViewedPage === 'trending'
+      ? trendingSorting
+      : lastViewedPage === 'bookmarked'
+        ? bookmarkSorting
+        : categorySorting
+
+  // States for navigation between projects
+  const [currentProjectIndex, setCurrentProjectIndex] = useState<number>()
+  const [previousProjectId, setPreviousProjectId] = useState<string>()
+  const [nextProjectId, setNextProjectId] = useState<string>()
+
+  // @TODO Add pagination
+  const [{ data: projectIds }] = useProjectIdsQuery({
+    variables: {
+      orderBy: sorting || defaultSort,
+      filter: {
+        ...filters
+      }
+    }
+  })
+
+  // List of all project IDs for navigation
+  const projects = projectIds?.projectCollection?.edges?.map(edge => edge.node) as Project[]
+
+  // Set IDs of previous and next project for navigation buttons
+  const updateProjectIndices = (currentId: string, projectList: Project[]) => {
+    const currentIndex = projectList.findIndex(p => p.id === currentId)
+
+    const newPreviousProjectId =
+      currentIndex > 0 ? (projectList[currentIndex - 1].id as string) : undefined
+
+    const newNextProjectId =
+      currentIndex < projectList.length - 1
+        ? (projectList[currentIndex + 1].id as string)
+        : undefined
+
+    setCurrentProjectIndex(currentIndex)
+    setPreviousProjectId(newPreviousProjectId)
+    setNextProjectId(newNextProjectId)
+  }
+
+  // Update project indices once projects are fetched
+  useEffect(() => {
+    if (projects && id) updateProjectIndices(id, projects)
+  }, [projects, id])
 
   return (
     <div className='flex h-[60px] w-full items-center gap-3 border-b border-white/5 bg-gray-900 px-4 lg:pl-7 lg:pr-3'>
@@ -63,7 +119,7 @@ const Navbar = ({
           <p className='text-sm text-white'>
             {currentProjectIndex !== undefined ? currentProjectIndex + 1 : '0'}&nbsp;
           </p>
-          <p className='text-sm text-white/50'>/&nbsp;{projectsLength}</p>
+          <p className='text-sm text-white/50'>/&nbsp;{projects?.length}</p>
         </div>
       )}
     </div>
